Add custom panic hook
Add a custom panic hook to shut down ncurses before printing the error.
The hook will exit ncurses mode before printing to stderr. This is not
fool proof in case a background thread crashes, but acceptable anyway:,
almost all work is done on the main thread and the hook is only intended
to help in manual debugging.

Because PanicInfo::message is currently nightly-only, rely on "{:#?}"
style formatting of the entire object.

Also, change UserInterface::shutdown from an instance method to a static
method to allow it to be called from anywhere.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>