Batch screen refreshes (16 ms refresh rate)
Use a combination of ncurses' 'wnoutrefresh' and 'doupdate' to limit the
number of screen refreshes: this noticeably cuts down on flickering.
Batch screen refreshes by scheduling a Refresh event 16 milliseconds
after the UI has changed. Any other change that happens before the time
is up will automatically be part of the scheduled refresh.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>